Track context canceled error in logger hook
