+++ pkg/auth/webapp_request_middleware.go
@@ -63,9 +63,6 @@ func (m *WebAppRequestMiddleware) Handle(next http.Handler) http.Handler {
 				m.TemplateEngine.RenderStatus(w, r, http.StatusNotFound, TemplateWebAppNotFoundHTML, data)
 			} else {
 				err = errorutil.ForceLogging(err)
-
-				otelauthgear.TrackContextCanceled(r.Context(), err)
-
 				// Our logging mechanism is not context-aware.
 				// We explicitly attach context here because it was the position we observed the log.
 				logger.WithContext(r.Context()).WithError(err).Error("failed to resolve config")
