+++ pkg/util/log/factory.go
@@ -21,6 +21,7 @@ func NewFactory(level Level, hooks ...logrus.Hook) *Factory {
 	logger.Level = level.Logrus()
 	logger.Out = ioutil.Discard
 	logger.Hooks.Add(&StackHook{})
+	logger.Hooks.Add(&TrackContextCanceledHook{})
 	logger.Hooks.Add(&ContextCauseHook{})
 	logger.Hooks.Add(&SkipLoggingHook{})
 	logger.Hooks.Add(NewWriterHook(os.Stderr))
